requirements.txt
tabula-py
JPype1